+++ lizard_layers/management/commands/sync_layers.py
@@ -22,12 +22,14 @@ class Command(BaseCommand):
     """
     Synchronizes ESF and EKR in datasources into cache table.
     """
-    def update_area_value(self, area, value, value_type):
+    def update_area_value(self, area, value, value_type, flag=None, comment=None):
         area_value, created = AreaValue.objects.get_or_create(
             area=area,
             value_type=value_type,
         )
         area_value.value = value
+        area_value.flag = flag
+        area_value.comment = comment
         area_value.save()
         return created
 
@@ -69,11 +71,11 @@ class Command(BaseCommand):
             return -1
         if len([j for j in judgements if j < 1]) > 1:
             return -2
-        
+
     def sync_ekr(self):
         value_type_worst = ValueType.objects.get(name='EKR-ONGUNSTIG')
         value_type_score = ValueType.objects.get(name='EKR-DOELSCORE')
-       
+
         parameter_types = list(ParameterType.objects.all())
 
         for area in Area.objects.filter(
@@ -86,23 +88,32 @@ class Command(BaseCommand):
                     parametercache=parameter_type.parameter,
                     geolocationcache__ident=area.ident,
                 )
+                value = None
+                flag = None
+                comment = None
                 if timeseries:
                     timeserie = timeseries[0]
                     try:
-                        value = timeserie.get_latest_event().value
+                        event = timeserie.get_latest_event(with_comments=True)
+                        value = event.value
+                        flag = event.flag
+                        comment = event.comment
                     except IndexError:
-                        value = None
-                else: value = None
+                        # No events at all
+                        pass
+
                 self.update_area_value(
                     area=area,
                     value=value,
+                    flag=flag,
+                    comment=comment,
                     value_type=parameter_type.value_type,
                 )
                 values.append({
                     'parameter_type': parameter_type,
                     'value': value
                 })
-            
+
             try:
                 value_worst = min([v['value'] for v in values
                                    if v['value'] is not None])